refactor(web): implement appeals view model
