+++ apps/web/src/app/views/home/controller.js
@@ -11,23 +11,8 @@ import { parse as parseUrl } from 'url';
 import { normalizeFilters, validateFilters } from '@pins/inspector-programming-lib/util/filtering.js';
 import { addSessionData, readSessionData } from '@pins/inspector-programming-lib/util/session.js';
 import { formatDateForDisplay } from '@pins/inspector-programming-lib/util/date.js';
-import { calendarViewModel } from './view-model.js';
+import { appealsViewModel, calendarViewModel } from './view-model.js';
 
-/**
- * @typedef {Object} Case
- * @property {string} caseId - The unique identifier for the case.
- * @property {string} caseType - The type of the case (e.g., 'W').
- * @property {string} caseProcedure - The procedure for the case (e.g., 'Written').
- * @property {string} allocationBand - The allocation band for the case.
- * @property {string} caseLevel - The level of the case.
- * @property {string} siteAddressPostcode - The postcode of the site address.
- * @property {string} lpaName - The name of the Local Planning Authority (LPA).
- * @property {string} lpaRegion - The region of the Local Planning Authority (LPA).
- * @property {string} caseStatus - The status of the case
- * @property {number} caseAge - The age of the case in weeks.
- * @property {number} linkedCases - The number of linked cases.
- * @property {Date} finalCommentsDate - The date of the final comments.
- */
 /**
  * @param {import('#service').WebService} service
  * @returns {import('express').Handler}
@@ -93,7 +78,8 @@ export function buildViewHome(service) {
 				specialisms,
 				pagination: paginationDetails,
 				query: {}
-			}
+			},
+			appeals: appealsViewModel(cases)
 		};
 
 		/**
@@ -118,8 +104,8 @@ export function buildViewHome(service) {
 			}
 		} else if (['calendar', 'inspector'].includes(req.query.currentTab)) {
 			calendarError = '';
-			inspectorError = 'Select an inspector';
-			errorSummary.push({
+      inspectorError = 'Select an inspector';
+			viewModel.errorSummary.push({
 				text: inspectorError,
 				href: '#inspectors'
 			});
@@ -132,7 +118,6 @@ export function buildViewHome(service) {
 
 		return res.render('views/home/view.njk', {
 			...viewModel,
-			cases: cases.map(caseViewModel),
 			inspectors,
 			data: formData,
 			apiKey: service.osMapsApiKey,
@@ -146,45 +131,11 @@ export function buildViewHome(service) {
 			specialisms,
 			specialismTypes: allocationLevels,
 			caseTypes,
-			inspectorError
+      inspectorError
 		});
 	};
 }
 
-export function getCaseColor(caseAge) {
-	if (caseAge > 40) return 'd4351c'; // red (41+ weeks)
-	if (caseAge > 20) return 'f47738'; // orange (21-40 weeks)
-	return '00703c'; // green (0-20 weeks)
-}
-
-/**
- *
- * @param {import('@pins/inspector-programming-lib/data/types.js').CaseViewModel[]} cases
- * @param {string} sort - The sort criteria, can be 'distance', 'hybrid', or 'age'.
- * @returns
- */
-export function sortCases(cases, sort) {
-	switch (sort) {
-		case 'distance':
-			//WIP
-			return cases;
-		case 'hybrid':
-			//WIP
-			return cases;
-		default:
-			return cases.sort((a, b) => b.caseAge - a.caseAge);
-	}
-}
-
-export function caseViewModel(c) {
-	return {
-		...c,
-		caseStatus: c.caseStatus?.replace('_', ' '),
-		finalCommentsDate: formatDateForDisplay(c.finalCommentsDate, { format: 'dd/MM/yyyy' }),
-		caseAgeColor: getCaseColor(c.caseAge),
-		currentDate: formatDateForDisplay(new Date(), { format: 'dd/MM/yyyy' })
-	};
-}
 export function buildPostHome(service) {
 	return async (req, res) => {
 		service.logger.info('post home');
