+++ apps/web/src/app/views/home/view.njk
@@ -288,7 +288,7 @@
             {% include 'views/home/cases-table.njk' %}
             </form>
 
-            {% if cases.length > 0 %}
+            {% if appeals.cases.length > 0 %}
                 <div class="govuk-form-group" style="display: flex; justify-content: space-between; align-items: center; margin: 0; padding: 0;">
                     <div style="margin: 0; padding: 0;">
                         {{ govukPagination(paginationDetails) }}
@@ -300,7 +300,7 @@
 
         {% set mapView %}
             <div class="home-map">
-                {{ map(apiKey, cases, inspectorPin) }}
+                {{ map(apiKey, appeals.cases, inspectorPin) }}
             </div>
         {% endset %}
